+++ src/interval.h
@@ -6,11 +6,11 @@
 class Interval
 {
 public:
-    double min, max;
+    float min, max;
 
     __host__ __device__ Interval() : min(RAY_INFINITY), max(-RAY_INFINITY) {} // Default interval is empty
 
-    __host__ __device__ Interval(double min, double max) : min(min), max(max) {}
+    __host__ __device__ Interval(float min, float max) : min(min), max(max) {}
 
     __host__ __device__ Interval(Interval a, Interval b)
     {
@@ -18,22 +18,22 @@ public:
         max = a.max > b.max ? a.max : b.max;
     }
 
-    __host__ __device__ double size() const
+    __host__ __device__ float size() const
     {
         return max - min;
     }
 
-    __host__ __device__ bool contains(double x) const
+    __host__ __device__ bool contains(float x) const
     {
         return min <= x && x <= max;
     }
 
-    __host__ __device__ bool surrounds(double x) const
+    __host__ __device__ bool surrounds(float x) const
     {
         return min < x && x < max;
     }
 
-    __host__ __device__ double clamp(double x) const
+    __host__ __device__ float clamp(float x) const
     {
         if (x < min)
             return min;
@@ -41,11 +41,11 @@ public:
             return max;
         return x;
     }
-
 };
 
-__host__ __device__ static Interval get_pixel_interval() {
-    return Interval(0.000, 0.999);
+__host__ __device__ static Interval get_pixel_interval()
+{
+    return Interval(0.000f, 0.999f);
 }
 
 #endif // INTERVAL_H_
