+++ src/sphere.h
@@ -7,12 +7,12 @@
 class Sphere : public hittable
 {
 public:
-    __device__ __host__ Sphere(const Vec3 &center, double radius, Material *mat_ptr) : center(center, Vec3(0, 0, 0)), radius(fmax(0.0, radius)), mat_ptr(mat_ptr)
+    __device__ __host__ Sphere(const Vec3 &center, float radius, Material *mat_ptr) : center(center, Vec3(0, 0, 0)), radius(fmax(0.0f, radius)), mat_ptr(mat_ptr)
     {
         auto rvec = Vec3(radius, radius, radius);
         box = aabb(center - rvec, center + rvec);
     }
-    Sphere(const Vec3 &center1, const Vec3 &center2, double radius, Material *mat) : center(center1, center2 - center1), radius(std::fmax(0, radius)), mat_ptr(mat)
+    Sphere(const Vec3 &center1, const Vec3 &center2, float radius, Material *mat) : center(center1, center2 - center1), radius(std::fmax(0, radius)), mat_ptr(mat)
     {
         auto rvec = Vec3(radius, radius, radius);
         aabb box1(center.at(0) - rvec, center.at(0) + rvec);
@@ -70,7 +70,7 @@ public:
 
 private:
     Ray center;
-    double radius;
+    float radius;
     Material *mat_ptr;
     aabb box;
 };
